# packages/api-contracts/openapi/backend.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/send-verification-code": {
      "post": {
        "operationId": "AuthController_sendVerificationCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVerificationCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "发送验证码（注册/密码重置）",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "用户注册",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "用户登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "申请重置密码",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/verify": {
      "post": {
        "operationId": "AuthController_verifyPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "验证重置密码验证码并更新密码",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "验证邮箱验证码",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "重新发送邮箱验证码",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/admin/auth/login": {
      "post": {
        "operationId": "AdminAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "管理员登录",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/admin/auth/refresh": {
      "post": {
        "operationId": "AdminAuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "刷新管理员访问令牌",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/admin/auth/register": {
      "post": {
        "operationId": "AdminAuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "注册首个管理员账号",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/admin/auth/me": {
      "get": {
        "operationId": "AdminAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取当前管理员资料",
        "tags": [
          "admin-auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Backend API",
    "description": "Auto generated OpenAPI for contracts",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user_123"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "user@example.com"
          },
          "nickname": {
            "type": "string",
            "description": "Nickname",
            "example": "John Doe",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar URL",
            "example": "https://example.com/avatar.png"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Whether email is verified",
            "example": true
          },
          "isGuest": {
            "type": "boolean",
            "description": "Whether the user is a guest account",
            "example": false
          },
          "roles": {
            "description": "User roles",
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "User creation time",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update time",
            "example": "2025-01-02T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "isGuest",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问令牌",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "description": "用户信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserProfileResponseDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "SendVerificationCodeRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "user@example.com"
          },
          "purpose": {
            "type": "string",
            "description": "Purpose of verification code",
            "enum": [
              "EMAIL_VERIFICATION",
              "PASSWORD_RESET"
            ],
            "example": "EMAIL_VERIFICATION"
          }
        },
        "required": [
          "email",
          "purpose"
        ]
      },
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "P@ssw0rd123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyPasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "6-digit reset code",
            "example": "654321"
          },
          "newPassword": {
            "type": "string",
            "description": "New password",
            "example": "P@ssw0rd123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
      "VerifyEmailRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "6-digit verification code",
            "example": "123456"
          },
          "updateUserStatus": {
            "type": "boolean",
            "description": "Whether to update user status automatically",
            "example": true,
            "default": true
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ResendVerificationRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "管理员登录用户名",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "管理员登录密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AdminProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "管理员 ID"
          },
          "username": {
            "type": "string",
            "description": "登录用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "nullable": true
          },
          "nickName": {
            "type": "string",
            "description": "昵称",
            "nullable": true
          },
          "isFrozen": {
            "type": "boolean",
            "description": "账号是否冻结"
          },
          "menuPermissions": {
            "description": "拥有的菜单权限编码",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "username",
          "isFrozen",
          "menuPermissions"
        ]
      },
      "AdminAuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问令牌"
          },
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          },
          "expiresIn": {
            "type": "string",
            "description": "访问令牌过期时间（字符串，例如 30m）"
          },
          "admin": {
            "description": "管理员信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminProfileDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "admin"
        ]
      },
      "AdminRefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AdminRegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "StrongP@ssw0rd"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "nickName": {
            "type": "string",
            "description": "昵称"
          },
          "roleCodes": {
            "description": "初始角色编码列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
    }
  }
}